global data and site data loading again

+++ website/src/ssg/renderer.py
@@ -7,50 +7,63 @@ from typing import Any, Coroutine, Iterator
 
 import aioshutil
 from aiopath import AsyncPath
+import jinja2
 
-from .config import Config
 from .data_directory import extract_global_data, text_to_data
 from .templates.jinja_renderer import build_jinja_environment
 
 
 
-
-async def run_render_pipeline(config: Config):
-    global_data = extract_global_data(base_path=config.global_data_dir)
-    renderer = JinjaRenderer.from_path(templates_dir=['site/templates'])
-    site_data = await read_and_render_dir(base_dir=config.site_data_dir, renderer=renderer, data=global_data)
-
+async def run_render_pipeline():
     
-    for renderfile_path in config.pages_dir.glob('[!_]*/_render.yaml'):
-        big_r = await TemplateRendering.from_file(renderfile_path, renderer, data=global_data)
+    # Copy Static Files
+    await asyncio.gather(
+        copy('site/static', '_output/static'),
+        copy('theme/assets', '_output/assets'),
+    )
 
-        for static_dir in renderfile_path.parent.glob(config.page_static_dirname):
-            await copy(static_dir, config.output_static_dir)
+    # Read Shared Data (no templating allowed)
+    global_data = extract_global_data(base_path='data')
+    
+    # Read site-wide data
+    env = build_jinja_environment()
+    site_data = await read_and_render_yaml_dir(base_dir='site/data', env=env, data=global_data)
+    breakpoint()
+    
+    
+    # for renderfile_path in config.pages_dir.glob('[!_]*/_render.yaml'):
+    #     big_r = await TemplateRendering.from_file(renderfile_path, renderer, data=global_data)
 
-        for page_render_data in big_r.page_instructions:
-            page_data = {}
-            for name, rel_path in page_render_data.page_data_files.items():
-                data_path = renderfile_path.parent.joinpath(rel_path)
-                data = await read_and_render_page_data(data_path, renderer, data=global_data)
-                page_data[name] = data
+    #     for static_dir in renderfile_path.parent.glob(config.page_static_dirname):
+    #         await copy(static_dir, config.output_static_dir)
 
-            page_html = await renderer.render_named_template(
-                template_path=renderfile_path.parent.joinpath(page_render_data.template),
-                **dict(
-                    data=global_data,
-                    site=site_data,
-                    page=page_data | page_render_data.data | {'url': page_render_data.url}
-                )
-            )
+    #     for page_render_data in big_r.page_instructions:
+    #         page_data = {}
+    #         for name, rel_path in page_render_data.page_data_files.items():
+    #             data_path = renderfile_path.parent.joinpath(rel_path)
+    #             data = await read_and_render_page_data(data_path, renderer, data=global_data)
+    #             page_data[name] = data
 
-            url_path = Path('./_output').joinpath(page_render_data.url.lstrip('/'))
-            await write_textfile(path=url_path, text=page_html)
+    #         page_html = await renderer.render_named_template(
+    #             template_path=renderfile_path.parent.joinpath(page_render_data.template),
+    #             **dict(
+    #                 data=global_data,
+    #                 site=site_data,
+    #                 page=page_data | page_render_data.data | {'url': page_render_data.url}
+    #             )
+    #         )
+
+    #         url_path = Path('./_output').joinpath(page_render_data.url.lstrip('/'))
+    #         await write_textfile(path=url_path, text=page_html)
 
 
-async def read_and_render_dir(base_dir: Path, renderer: JinjaRenderer, **render_data):
+async def read_and_render_yaml_dir(base_dir: str | Path, env: jinja2.Environment, **render_data):
     data = {}
     async for path in AsyncPath(base_dir).glob('*.yaml'):
-        data[path.stem] = await read_yaml(path, renderer, **render_data)
+        text = await AsyncPath(path).read_text()
+        template = env.from_string(text)
+        text_to_load = await template.render_async(**render_data)
+        data[path.stem] = text_to_data(text_to_load, format='yaml')
     return data
 
 
@@ -63,12 +76,6 @@ async def read_and_render_page_data(path, renderer: JinjaRenderer, **render_data
             
 
 
-def copy_static_files(config: Config, skip_if_exists: bool = False) -> Iterator[Coroutine]:
-    for src, target in config.static_path_map.items():
-        if skip_if_exists and target.exists():
-            continue
-        yield copy(src, target)
-
 
 ####### UTILS #####################
 
@@ -79,16 +86,13 @@ async def write_textfile(path, text) -> None:
     await apath.write_text(text)
 
 
-async def read_yaml(path: Path, renderer: JinjaRenderer = None,  **render_data):
-    text = await AsyncPath(path).read_text()
-    text_to_load = await renderer.render_in_place(template_text=text, **render_data) if renderer is not None else text
-    data = text_to_data(text_to_load, format='yaml')
-    return data
 
 
 
-async def copy(src: Path, target: Path) -> None:
+async def copy(src: Path, target: Path, skip_if_exists: bool = True) -> None:
     print(f"Copying: {src}")
+    if skip_if_exists and Path(target).exists():
+        return
     if Path(src).is_dir():
         await aioshutil.copytree(src, target, dirs_exist_ok=True)
         return
@@ -98,3 +102,4 @@ async def copy(src: Path, target: Path) -> None:
 
 
         
+
